refactor: create custom hooks for insert batch and reorder list item mutations
remove unnecessary prop drilling from `DraggableItems` component
