+++ client/src/api/mutation-hooks/list-item/index.ts
@@ -1,4 +1,6 @@
 export * from "./useAppendListItemMutation";
 export * from "./useDeleteListItemMutation";
+export * from "./useInsertBatchListItemsMutation";
 export * from "./useInsertListItemMutation";
+export * from "./useReorderListItemsMutation";
 export * from "./useUpdateListItemMutation";
